Move formatPEM() into examples folder
