+++ examples/CRLComplexExample/es6.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-unused-vars,no-undef,no-unreachable */
 import { arrayBufferToString, bufferToHexCodes, toBase64 } from "pvutils";
-import { getCrypto, getAlgorithmParameters, setEngine, formatPEM } from "../../src/common";
+import { getCrypto, getAlgorithmParameters, setEngine } from "../../src/common";
+import { formatPEM } from "../examples_common";
 import * as asn1js from "asn1js";
 import Certificate from "../../src/Certificate";
 import CertificateRevocationList from "../../src/CertificateRevocationList";
