+++ examples/HowToEncryptCMSviaPassword/es6.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-undef,no-unreachable */
 import * as asn1js from "asn1js";
 import { arrayBufferToString, stringToArrayBuffer, fromBase64, toBase64, isEqualBuffer } from "pvutils";
-import { setEngine, formatPEM } from "../../src/common";
+import { setEngine } from "../../src/common";
+import { formatPEM } from "../examples_common";
 import EnvelopedData from "../../src/EnvelopedData";
 import ContentInfo from "../../src/ContentInfo";
 //<nodewebcryptoossl>
