+++ examples/OCSPRequestComplexExample/es6.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-undef,no-unreachable,no-unused-vars */
 import * as asn1js from "asn1js";
 import { bufferToHexCodes, toBase64,arrayBufferToString } from "pvutils";
-import { setEngine, formatPEM } from "../../src/common";
+import { setEngine } from "../../src/common";
+import { formatPEM } from "../examples_common";
 import OCSPRequest from "../../src/OCSPRequest";
 import GeneralName from "../../src/GeneralName";
 import RelativeDistinguishedNames from "../../src/RelativeDistinguishedNames";
