+++ examples/P7BSimpleExample/es6.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-undef,no-unreachable */
 import * as asn1js from "asn1js";
-import { getCrypto, getAlgorithmParameters, setEngine, formatPEM } from "../../src/common";
+import { getCrypto, getAlgorithmParameters, setEngine } from "../../src/common";
+import { formatPEM } from "../examples_common";
 import { arrayBufferToString, toBase64 } from "pvutils";
 import Certificate from "../../src/Certificate";
 import AttributeTypeAndValue from "../../src/AttributeTypeAndValue";
