+++ examples/PKCS10ComplexExample/es6.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-undef,no-unreachable */
 import * as asn1js from "asn1js";
 import { arrayBufferToString, stringToArrayBuffer, toBase64, fromBase64 } from "pvutils";
-import { getCrypto, getAlgorithmParameters, setEngine, formatPEM } from "../../src/common.js";
+import { getCrypto, getAlgorithmParameters, setEngine } from "../../src/common.js";
+import { formatPEM } from "../examples_common.js";
 import CertificationRequest from "../../src/CertificationRequest.js";
 import AttributeTypeAndValue from "../../src/AttributeTypeAndValue.js";
 import Attribute from "../../src/Attribute.js";
