+++ examples/PKCS12SimpleExample/es6.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-undef,no-unreachable */
 import * as asn1js from "asn1js";
 import { stringToArrayBuffer, arrayBufferToString, fromBase64, toBase64 } from "pvutils";
-import { getCrypto, getAlgorithmParameters, getRandomValues, setEngine, formatPEM } from "../../src/common";
+import { getCrypto, getAlgorithmParameters, getRandomValues, setEngine } from "../../src/common";
+import { formatPEM } from "../examples_common";
 import Certificate from "../../src/Certificate";
 import PrivateKeyInfo from "../../src/PrivateKeyInfo";
 import AuthenticatedSafe from "../../src/AuthenticatedSafe";
