+++ examples/SMIMEEncryptionExample/es6.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-undef,no-unreachable */
 import * as asn1js from "asn1js";
 import { arrayBufferToString, stringToArrayBuffer } from "pvutils";
-import { getCrypto, getAlgorithmParameters, formatPEM } from "../../src/common";
+import { getCrypto, getAlgorithmParameters } from "../../src/common";
+import { formatPEM } from "../examples_common";
 import Certificate from "../../src/Certificate";
 import EnvelopedData from "../../src/EnvelopedData";
 import ContentInfo from "../../src/ContentInfo";
