+++ examples/TSPResponseComplexExample/es6.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-undef,no-unreachable,no-unused-vars */
 import * as asn1js from "asn1js";
 import { stringToArrayBuffer, arrayBufferToString, bufferToHexCodes, toBase64 } from "pvutils";
-import { getCrypto, getAlgorithmParameters, setEngine, getOIDByAlgorithm, formatPEM } from "../../src/common";
+import { getCrypto, getAlgorithmParameters, setEngine, getOIDByAlgorithm } from "../../src/common";
+import { formatPEM } from "../examples_common";
 import Certificate from "../../src/Certificate";
 import AttributeTypeAndValue from "../../src/AttributeTypeAndValue";
 import Extension from "../../src/Extension";
